requirements.txt
django
requests
psycopg2
djangorestframework
django-filter
djangorestframework-gis
django-cors-headers
